feat(cheatcodes): extract crypto cheats into their own category (#8578)
